refactor: move services to dedicated folder
