+++ src/app/services/equipment.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
-
 import { EquipmentService } from './equipment.service';
+
 
 describe('EquipmentService', () => {
   let service: EquipmentService;
